refactor: standardize loading states and split onboarding into page/container/ui
- Add shared LoadingSpinner and PageLoading components (loading-spinner.tsx)
- Replace ad-hoc loading UIs (text, MUI CircularProgress, return null) with PageLoading across profile, reservation, email-verified, and edit pages
- Add Loader2 spinner to AuthButton when isSubmitting
- Replace MUI CircularProgress in mentor-pool/ui.tsx with LoadingSpinner
- Split onboarding page into container (logic/state) and ui (presentation)
- Replace MUI ArrowBackIcon with lucide-react ChevronLeft in onboarding
- Add isSubmitting loading state with spinner on onboarding final submit button
- Remove console.error from onboarding submit handler
