+++ src/app/reservation/mentor/container.tsx
@@ -2,6 +2,7 @@
 
 import dynamic from 'next/dynamic';
 
+import { PageLoading } from '@/components/ui/loading-spinner';
 import { useReservationData } from '@/hooks/user/reservation/useReservationData';
 
 const ReservationPresentation = dynamic(() => import('./ui'));
@@ -9,6 +10,6 @@ const ReservationPresentation = dynamic(() => import('./ui'));
 export default function ReservationContainer() {
   const { data, isLoading } = useReservationData();
 
-  if (isLoading || !data) return <div className="p-6">Loading…</div>;
+  if (isLoading || !data) return <PageLoading />;
   return <ReservationPresentation {...data} />;
 }
