+++ src/components/ui/loading-spinner.tsx
@@ -1,0 +1,37 @@
+import { Loader2 } from 'lucide-react';
+
+import { cn } from '@/lib/utils';
+
+interface LoadingSpinnerProps {
+  className?: string;
+  size?: 'sm' | 'md' | 'lg';
+}
+
+const sizeClasses = {
+  sm: 'h-4 w-4',
+  md: 'h-6 w-6',
+  lg: 'h-8 w-8',
+};
+
+export function LoadingSpinner({
+  className,
+  size = 'md',
+}: LoadingSpinnerProps) {
+  return (
+    <Loader2
+      className={cn(
+        'animate-spin text-muted-foreground',
+        sizeClasses[size],
+        className
+      )}
+    />
+  );
+}
+
+export function PageLoading() {
+  return (
+    <div className="flex h-[50vh] w-full items-center justify-center">
+      <LoadingSpinner size="lg" />
+    </div>
+  );
+}
